+++ package.json
@@ -25,7 +25,7 @@
     "bootstrap": "^4.6.0",
     "gh-pages": "^3.1.0",
     "js-cookie": "^2.2.1",
-    "node-sass": "^5.0.0",
+    "node-sass": "^8.0.0",
     "react-bootstrap": "^1.5.2",
     "react-color": "^2.19.3",
     "react-google-login": "^5.2.2",
