+++ src/components/PlanDeEstudio/BarrasDeProgreso/BarrasDeProgreso.js
