+++ src/components/PlanDeEstudio/BarrasDeProgreso/BarrasDeProgreso.test.js
