+++ src/components/PlanDeEstudio/BotonesDeColor/BotonesDeColor.js
