+++ src/components/PlanDeEstudio/BotonesDeColor/BotonesDeColor.test.js
