+++ src/components/PlanDeEstudio/PlanDeEstudio.js
